Extract Wikipedia plugin into a separate file
