+++ plugins/__init__.py
@@ -7,4 +7,4 @@ __all__ = ['plugins', 'command_catcher', 'first_plugin', 'auto_join', 'aduno',
             'tenta', 'prisjakt', 'spotify', 'stava', 'stock', 'down', 'metacritic',
             'notes', 'fml', 'systembolaget', 'randombuy', 'festern_bbq',
             'compliment', 'roulette', 'tyda', 'yrno', 'tweet', 'calc', 'degoogle',
-            'temperatur_nu', 'google_search']
+            'temperatur_nu', 'google_search', 'wikipedia']
